+++ packages/amplify-category-api/src/commands/api/add-graphql-datasource.js
@@ -114,8 +114,17 @@ module.exports = {
         const schemaDirectoryExists = fs.existsSync(schemaDirectoryPath);
 
         if (schemaFileExists) {
+          const typesToBeMerged = [rdsGraphQLSchemaDoc];
           const currGraphQLSchemaDoc = readSchema(graphqlSchemaFilePath);
-          const concatGraphQLSchemaDoc = mergeTypeDefs([currGraphQLSchemaDoc, rdsGraphQLSchemaDoc], { all: true });
+
+          if (currGraphQLSchemaDoc) {
+            typesToBeMerged.unshift(currGraphQLSchemaDoc);
+          } else {
+            context.print.warning(`Graphql Schema file "${graphqlSchemaFilePath}" is empty.`);
+            context.print.info('');
+          }
+
+          const concatGraphQLSchemaDoc = mergeTypeDefs(typesToBeMerged, { all: true });
           fs.writeFileSync(graphqlSchemaFilePath, graphql.print(concatGraphQLSchemaDoc), 'utf8');
         } else if (schemaDirectoryExists) {
           const rdsSchemaFilePath = path.join(schemaDirectoryPath, 'rds.graphql');
@@ -222,6 +231,11 @@ async function getAwsClient(context, action) {
 
 function readSchema(graphqlSchemaFilePath) {
   const graphqlSchemaRaw = fs.readFileSync(graphqlSchemaFilePath).toString();
+  const graphqlSchemaIsEmpty = graphqlSchemaRaw.trim().length === 0;
+  if (graphqlSchemaIsEmpty) {
+    return null;
+  }
+
   let currGraphQLSchemaDoc;
   try {
     currGraphQLSchemaDoc = graphql.parse(graphqlSchemaRaw);