+++ packages/amplify-category-auth/provider-utils/awscloudformation/triggers/PreTokenGeneration/function-template-dir/trigger-index.js
@@ -3,11 +3,12 @@
   provided that the file names (without extension) are included in the "MODULES" env variable.
   "MODULES" is a comma-delimmited string.
 */
+const moduleNames = process.env.MODULES.split(',');
+const modules = moduleNames.map(name => require(`./${name}`));
 
 exports.handler = (event, context, callback) => {
-  const modules = process.env.MODULES.split(',');
   for (let i = 0; i < modules.length; i += 1) {
-    const { handler } = require(`./${modules[i]}`);
+    const { handler } = modules[i];
     handler(event, context, callback);
   }
 };