+++ packages/amplify-frontend-ios/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.18.11](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-ios@2.18.10...amplify-frontend-ios@2.18.11) (2021-03-11)
+
+**Note:** Version bump only for package amplify-frontend-ios
+
+
+
+
+
 ## [2.18.10](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-ios@2.18.9...amplify-frontend-ios@2.18.10) (2021-03-05)
 
 **Note:** Version bump only for package amplify-frontend-ios