+++ packages/amplify-python-function-runtime-provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-python-function-runtime-provider@1.6.3...amplify-python-function-runtime-provider@1.6.4) (2021-03-11)
+
+**Note:** Version bump only for package amplify-python-function-runtime-provider
+
+
+
+
+
 ## [1.6.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-python-function-runtime-provider@1.6.2...amplify-python-function-runtime-provider@1.6.3) (2021-03-05)
 
 **Note:** Version bump only for package amplify-python-function-runtime-provider